Swapped to dual color checking
Before we'd check the combined mask to see, but certain objects are the right shades of yellow (olmec) so that we get a false positive.  It never should matter unless you tried to make that happen as it would not be positive long enough to run the timer.  However, due to this we analyze each channel separately now; thresholds are still being worked out.